Renamed Debug namespace into Diagnostics
