+++ src/ChainSafe.Gaming/Diagnostics/AddressExtensions.cs
@@ -1,16 +1,12 @@
-namespace ChainSafe.Gaming.Debug
+using ChainSafe.Gaming.Utils;
+
+namespace ChainSafe.Gaming.Diagnostics
 {
     public static class AddressExtensions
     {
-        public static bool IsPublicAddress(string value)
-        {
-            // todo: more accurate test
-            return value.Length == 42;
-        }
-
         public static string AssertIsPublicAddress(this string value, string variableName)
         {
-            if (!IsPublicAddress(value))
+            if (!AddressUtil.IsPublicAddress(value))
             {
                 throw new Web3Exception($"\"{variableName}\" is not public address");
             }
